BUG FIX: Database and Users
I recreated the database so that now the auctions_user table is created. Turns out that was was producing the faulty database was that the 'makemigrations' command was not detecting any of the models in the models.py file. To fix that, I needed to delete the database, and then use the command 'makemigrations auctions'.