Created Listings Model (BUGGY)
I was able to successfully create the Listing model, and most of the fields for the auctions_listing table were correctly created. However, the field 'created_on', which stores the date and time when the listing was created, is not being created, so it's not being added to the database. Find out why.