Updated 'Create Listing' Algorithm
I finally added the code in the create() view so that the user can now fill in the form in the /create page, and then send that data into the database. However, I still need to add some more data into some fields of the Listing table such as the date in which the listing was created, the category, and the URL of the picture.