Added Category Input (BUGGY)
The Create Listing page is showing the 'Category' input as a dropdown menu. The problem is that the menu is displaying 'Object category (1/2/3/4)' instead of the category names, and, when I choose one, the resulting entry will have the ID for that category for the new listing, instead of saying 'gaming' or any category name.