Updated Listing Page's Algorithm
The display_listing() view can now check if the user placed a bid lower than the initial price, if the user is the first bidder or not, and if the current bid is higher than any bid placed by anyone else. Also, if the user's the first one to bid, they get a message telling them that they are able to plce a bid that's the same as the price beign displayed. Now, if another user has bidded before, the current user will only be able to place bids that are higher than the previous bid (not equal).